git: add script that finds undeleted branches of closed github PRs
The python script queries all PRs from the github API for a repository.
If for the associated branch only closed PRs exist, the branch is not
deleted and the branch is neither "master" nor develop, it's printed to
STDOUT.
The list of branches is also written to a result file at the end.
Status messages are printed to STDERR.

The output of the script can be used to delete all obsolete PR branches of
a github repository.